Move HTTP binding helpers into plugins service
Interface diff:

- import "github.com/valon-technologies/gestalt/server/internal/httpbinding"

+ import "github.com/valon-technologies/gestalt/server/services/plugins/httpbinding"

Examples:

httpbinding.ValidateHTTPSecurityScheme(path, scheme)

httpbinding.RenderPayloadTemplate(template, headers, rawBody)
