+++ gestaltd/services/plugins/httpbinding/payload_template.go
